Update NewToki Extension to v1.2.19 (#5258)
* Mitigation about IP Ban by bunch of request on latest on manatoki.

RateLimit, And make default to not make bunch of request.
Split ManaToki into separate file as it became bigger.

* More rate limit

NewToki and ManaToki Shares IP bans.
Also when refresh the batch of mangas, It could be banned too.

So rate limit on ChapterList and MangaDetail request too. (And apply both of them)

* Fix lint about `java.util` and some format